Moving the page placeholder models over to title
